refactor(node): replace 4 Fs*Repo with unified FsStorage
Single FsStorage struct implements Storage directly. Holds all paths
and state that were split across FsMemoryRepo, FsSkillRepo,
FsSessionRepo, and FsAgentRepo. NodeRepos composite and the old
sub-repo files deleted.
